+++ lmfdb/artin_representations/databases/Dokchitser_databases.py
