+++ lmfdb/artin_representations/databases/Lfunction_databases.py
