+++ lmfdb/artin_representations/databases/standard_types.py
