+++ lmfdb/artin_representations/databases/type_generation.py
